% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparam_link_analysis.R
\name{nonparam_link_analysis}
\alias{nonparam_link_analysis}
\title{Non-parametric link analysis between two samples}
\usage{
nonparam_link_analysis(sample1, sample2, type1, type2)
}
\arguments{
\item{sample1}{The first sample}

\item{sample2}{The second sample}

\item{type1}{the data type of first sample (can be interval, ordinal and nominal)}

\item{type2}{the data type of second sample (can be interval, ordinal and nominal)}
}
\value{
The score of a linkage (mostly p-value, can be also a z-score in Mann-Whitney Test, or correlation coefficient in Spearman Correlation Test)
}
\description{
Check for any kind of linkage between two unrelated samples
}
\examples{
result <- nonparam_link_analysis(c(1, 2, 3, 4, 5), c(6, 7, 8, 9, 10), "int", "ord")
result <- nonparam_link_analysis(c("A", "A", "B", "C"), c(58, 70, 64, 19), "ord", "nom")
}
